chore: keyword, news - model fixed and repository refactoring
